requirements.txt
# Core packages
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
seaborn==0.12.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3

# Technical analysis
mplfinance==0.12.9b7
ta-lib==0.4.26

# Visualization
plotly==5.14.1
kaleido==0.2.1

# Machine learning (for predictions)
scikit-learn==1.2.2
tensorflow==2.13.0
keras==2.13.1

# API integrations
websocket-client==1.5.1
ccxt==3.1.43  # Crypto exchange library

# Utilities
tqdm==4.65.0
loguru==0.7.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.3.0